box button type, size 분리
