+++ sdg/src/main/java/com/shopl/sdg/component/button/box/SDGBoxButton.kt
@@ -41,49 +41,6 @@ import com.shopl.sdg_common.ui.components.IOText
 import com.shopl.sdg_common.ui.components.IOTypeface
 import com.shopl.sdg_resource.R
 
-sealed class SDGBoxButtonSize(
-    val height: Dp,
-    val horizontalPadding: Dp,
-    val gap: Dp,
-    val radius: Dp,
-    val labelSize: TextUnit,
-) {
-    @Stable
-    object Medium : SDGBoxButtonSize(
-        height = 42.dp,
-        horizontalPadding = 16.dp,
-        gap = 4.dp,
-        radius = 12.dp,
-        labelSize = 14.sp,
-    )
-
-    @Stable
-    object Small : SDGBoxButtonSize(
-        height = 32.dp,
-        horizontalPadding = 10.dp,
-        gap = 2.dp,
-        radius = 8.dp,
-        labelSize = 12.sp,
-    )
-
-    @Stable
-    object XSmall : SDGBoxButtonSize(
-        height = 20.dp,
-        horizontalPadding = 6.dp,
-        gap = 2.dp,
-        radius = 6.dp,
-        labelSize = 12.sp,
-    )
-}
-
-sealed class SDGBoxButtonType {
-    @Stable
-    object Solid : SDGBoxButtonType()
-
-    @Stable
-    data class Line(val lineColor: Color) : SDGBoxButtonType()
-}
-
 /**
  * SDG - Button - Box Button
  *
